+++ script/structure_creation/create_new_structure.py
@@ -45,7 +45,7 @@ while restart:
     avoid_tags = ""
     avoid_tag = input("\nEnter structure tags to avoid. Type 'exit' to quit\n")
     while(avoid_tag != 'exit'): 
-        avoid_tags = avoid_tags + ", STRUCTURE_TAGS." + avoid_tags.upper().strip()
+        avoid_tags = avoid_tags + ", RSStructureTagMap.STRUCTURE_TAGS." + avoid_tags.upper().strip()
         avoid_tag = input()
     avoid_tags = avoid_tags[2:]
 
@@ -117,7 +117,7 @@ while restart:
         file.write(file_content)
 
     with open(os.path.join('template', 'fabric_biome_spawn.txt'), "r") as file:
-        file_content = file.read().replace("$1", structure_registry_name).replace("$2", config_subcategory).replace("$3", config_category) \
+        file_content = file.read().replace("$1", structure_registry_name).replace("$2", config_subcategory[:-1]).replace("$3", config_category) \
                                 .replace("$4", config_spawnrate_entry).replace("$5", config_modded_biome_entry).replace("$6", structure_variable_name)
     with open(os.path.join('code', 'fabric', structure_registry_name+'_biome_spawn.txt'), "w") as file:
         raw_output += "\n\n" + file_content
@@ -201,34 +201,34 @@ while restart:
         file.write(file_content)
 
 
-    raw_output += "\n\n--------------ADVANCEMENTS-------------"
+    #  --------------ADVANCEMENTS--------------
 
 
     with open(os.path.join('template', 'translated_advancement.json'), "r") as file:
-        file_content = file.read().replace("$1", structure_registry_name).replace("$2", config_subcategory).replace("$3", advancement_icon) \
-                                .replace("$3", advancement_title).replace("$3", advancement_description).replace("$3", advancement_exp)
-    with open(os.path.join('advancements', 'Repurposed_Structures-English_Translated_Advancements', "data", "repurposed_structures", "advancements", config_subcategory, structure_registry_name+'.json'), "w") as file:
+        file_content = file.read().replace("$1", structure_registry_name).replace("$2", advancement_icon) \
+                                .replace("$3", advancement_title).replace("$4", advancement_description).replace("$5", advancement_exp)
+    with open(os.path.join('advancements', 'Repurposed_Structures-English_Translated_Advancements', "data", "repurposed_structures", "advancements", config_subcategory[:-1], structure_registry_name+'.json'), "w") as file:
         file.write(file_content)
 
 
     with open(os.path.join('template', 'hidden_advancement.json'), "r") as file:
-        file_content = file.read().replace("$1", structure_registry_name).replace("$2", config_subcategory).replace("$3", advancement_icon) \
-                                .replace("$3", advancement_title).replace("$3", advancement_description).replace("$3", advancement_exp)
-    with open(os.path.join('advancements', 'Repurposed_Structures-Hidden_Advancements', "data", "repurposed_structures", "advancements", config_subcategory, structure_registry_name+'.json'), "w") as file:
+        file_content = file.read().replace("$1", structure_registry_name).replace("$2", advancement_icon) \
+                                .replace("$3", advancement_title).replace("$4", advancement_description).replace("$5", advancement_exp)
+    with open(os.path.join('advancements', 'Repurposed_Structures-Hidden_Advancements', "data", "repurposed_structures", "advancements", config_subcategory[:-1], structure_registry_name+'.json'), "w") as file:
         file.write(file_content)
 
 
     with open(os.path.join('template', 'translation_advancement.json'), "r") as file:
-        file_content = file.read().replace("$1", structure_registry_name).replace("$2", config_subcategory).replace("$3", advancement_icon) \
-                                .replace("$3", advancement_title).replace("$3", advancement_description).replace("$3", advancement_exp)
-    with open(os.path.join('advancements', 'Repurposed_Structures-Translation_Advancements', "data", "repurposed_structures", "advancements", config_subcategory, structure_registry_name+'.json'), "w") as file:
+        file_content = file.read().replace("$1", structure_registry_name).replace("$2", advancement_icon) \
+                                .replace("$3", advancement_title).replace("$4", advancement_description).replace("$5", advancement_exp)
+    with open(os.path.join('advancements', 'Repurposed_Structures-Translation_Advancements', "data", "repurposed_structures", "advancements", config_subcategory[:-1], structure_registry_name+'.json'), "w") as file:
         file.write(file_content)
 
 
     with open(os.path.join('template', 'disabled_advancement.json'), "r") as file:
-        file_content = file.read().replace("$1", structure_registry_name).replace("$2", config_subcategory).replace("$3", advancement_icon) \
-                                .replace("$3", advancement_title).replace("$3", advancement_description).replace("$3", advancement_exp)
-    with open(os.path.join('advancements', 'Repurposed_Structures-Disabled_Advancements', "data", "repurposed_structures", "advancements", config_subcategory, structure_registry_name+'.json'), "w") as file:
+        file_content = file.read().replace("$1", structure_registry_name).replace("$2", advancement_icon) \
+                                .replace("$3", advancement_title).replace("$4", advancement_description).replace("$5", advancement_exp)
+    with open(os.path.join('advancements', 'Repurposed_Structures-Disabled_Advancements', "data", "repurposed_structures", "advancements", config_subcategory[:-1], structure_registry_name+'.json'), "w") as file:
         file.write(file_content)